+++ packages/starters/auth/pb_migrations/1738527881_updated_users.js
@@ -1,24 +1,33 @@
 /// <reference path="../pb_data/types.d.ts" />
-migrate((app) => {
-  const collection = app.findCollectionByNameOrId("_pb_users_auth_")
+migrate(
+  (app) => {
+    const collection = app.findCollectionByNameOrId('_pb_users_auth_')
 
-  // update collection data
-  unmarshal({
-    "resetPasswordTemplate": {
-      "body": "<p>Hello,</p>\n<p>Click on the button below to reset your password.</p>\n<p>\n  <a class=\"btn\" href=\"{APP_URL}/auth/confirm/{TOKEN}/password-reset\" target=\"_blank\" rel=\"noopener\">Reset password</a>\n</p>\n<p><i>If you didn't ask to reset your password, you can ignore this email.</i></p>\n<p>\n  Thanks,<br/>\n  {APP_NAME} team\n</p>"
-    }
-  }, collection)
+    // update collection data
+    unmarshal(
+      {
+        resetPasswordTemplate: {
+          body: '<p>Hello,</p>\n<p>Click on the button below to reset your password.</p>\n<p>\n  <a class="btn" href="{APP_URL}/auth/confirm/{TOKEN}/password-reset" target="_blank" rel="noopener">Reset password</a>\n</p>\n<p><i>If you didn\'t ask to reset your password, you can ignore this email.</i></p>\n<p>\n  Thanks,<br/>\n  {APP_NAME} team\n</p>',
+        },
+      },
+      collection
+    )
 
-  return app.save(collection)
-}, (app) => {
-  const collection = app.findCollectionByNameOrId("_pb_users_auth_")
+    return app.save(collection)
+  },
+  (app) => {
+    const collection = app.findCollectionByNameOrId('_pb_users_auth_')
 
-  // update collection data
-  unmarshal({
-    "resetPasswordTemplate": {
-      "body": "<p>Hello,</p>\n<p>Click on the button below to reset your password.</p>\n<p>\n  <a class=\"btn\" href=\"{APP_URL}/auth/confirm-password-reset/{TOKEN}\" target=\"_blank\" rel=\"noopener\">Reset password</a>\n</p>\n<p><i>If you didn't ask to reset your password, you can ignore this email.</i></p>\n<p>\n  Thanks,<br/>\n  {APP_NAME} team\n</p>"
-    }
-  }, collection)
+    // update collection data
+    unmarshal(
+      {
+        resetPasswordTemplate: {
+          body: '<p>Hello,</p>\n<p>Click on the button below to reset your password.</p>\n<p>\n  <a class="btn" href="{APP_URL}/auth/confirm-password-reset/{TOKEN}" target="_blank" rel="noopener">Reset password</a>\n</p>\n<p><i>If you didn\'t ask to reset your password, you can ignore this email.</i></p>\n<p>\n  Thanks,<br/>\n  {APP_NAME} team\n</p>',
+        },
+      },
+      collection
+    )
 
-  return app.save(collection)
-})
+    return app.save(collection)
+  }
+)