+++ packages/starters/auth/pb_migrations/1738530766_updated_users.js
@@ -1,24 +1,33 @@
 /// <reference path="../pb_data/types.d.ts" />
-migrate((app) => {
-  const collection = app.findCollectionByNameOrId("_pb_users_auth_")
+migrate(
+  (app) => {
+    const collection = app.findCollectionByNameOrId('_pb_users_auth_')
 
-  // update collection data
-  unmarshal({
-    "otp": {
-      "enabled": true
-    }
-  }, collection)
+    // update collection data
+    unmarshal(
+      {
+        otp: {
+          enabled: true,
+        },
+      },
+      collection
+    )
 
-  return app.save(collection)
-}, (app) => {
-  const collection = app.findCollectionByNameOrId("_pb_users_auth_")
+    return app.save(collection)
+  },
+  (app) => {
+    const collection = app.findCollectionByNameOrId('_pb_users_auth_')
 
-  // update collection data
-  unmarshal({
-    "otp": {
-      "enabled": false
-    }
-  }, collection)
+    // update collection data
+    unmarshal(
+      {
+        otp: {
+          enabled: false,
+        },
+      },
+      collection
+    )
 
-  return app.save(collection)
-})
+    return app.save(collection)
+  }
+)