+++ packages/starters/datastar/pb_hooks/pages/+config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  plugins: [
+    'pocketpages-plugin-ejs',
+    'pocketpages-plugin-datastar',
+    'pocketpages-plugin-realtime',
+  ],
+}